const std::string for robot_ip_ member variable
Using const char *robot_ip_ gets messed up when there is a failure
to connect. It appears that the pointer points to something else
which results in the string being "usRaw". Using a std::string
fixes the problem and is safer anyway. This should be fixed
in other classes too