Overhaul waypoint actions (working)
The previous waypoint action implementation was working
but much of the logic was in Lua and hardcoded in
"preset" values for the ActionDoneCalc records for
open/close gripper. Now waypoint_actions10.db loads 10
configurable waypoint actions (FLNK + calcout). The
link and calcout are autosaved. The downside to this is
I haven't found a way to overwrite the ZRST, ONST, etc.
values in the WaypointJ:$(N):ActionOpt mbbo records,
so the choices for Actions on the waypoint screens is
just Action 1, Action 2, etc.