automatcally ResetJCmd and ResetPoseCmd
Now the user should not see the "Reset" PV/button at all.
The command values for the joint and pose "axes" are automatically
synced to the readbacks after motion completes, similar to the motor record

This required adding a dependency on the calc module in the IOC that
uses the rtde_control.db database because the sseq record is used.