Busy PV for each waypoint
Each waypoint now has a Busy PV which is "Busy" (1) when the robot is
en-route to the waypoint, otherwise it is "Done" (0). The Waypoint_10x
screens now show a yellow light next to each waypoint when the robot is
on the way to it. this shoudl be linked to a corresponding path waypoint
busy PV as well.